+++ src/core/Media.h
@@ -91,6 +91,27 @@ public:
     libvlc_media_t *core();
 
     /*!
+        \brief Get parsed status
+
+        Know when it is OK to read metadata and track information
+
+        \return parsed status (const bool)
+        \since VLC-Qt 1.1
+    */
+    bool parsed() const;
+
+    /*!
+        \brief Parse media information
+
+        Parse media information: metadata and track information.
+        This call is async.
+
+        \see parsedChanged
+        \since VLC-Qt 1.1
+    */
+    void parse();
+
+    /*!
         \brief Current media location
 
         \return QString media location
@@ -103,6 +124,20 @@ public:
         \return VlcStats media stats object
     */
     VlcStats *getStats();
+
+    /*!
+        \brief Get media state
+        \return current media state
+        \since VLC-Qt 1.1
+    */
+    Vlc::State state() const;
+
+    /*!
+        \brief Get media duration
+        \return duration
+        \since VLC-Qt 1.1
+    */
+    qint64 duration() const;
 
     /*!
         \brief Duplicate (provided for convenience)
@@ -283,8 +318,15 @@ signals:
     /*!
         \brief Signal sent on parsed change
         \param status new parsed status
+        \deprecated Deprecated since VLC-Qt 1.1, will be removed in 2.0
     */
-    void parsedChanged(int status);
+    Q_DECL_DEPRECATED void parsedChanged(int status);
+
+    /*!
+        \brief Signal sent on parsed change
+        \param status new parsed status
+    */
+    void parsedChanged(bool status);
 
     /*!
         \brief Signal sent on freed